movistar-u7d.py & u7d.py: EUREKA, we found how to signal a stream has ended, turns out the new sanic streaming api has two parameters in the .send() method, the second one being a bool, end_stream, so just by doing:
    await resp.send(b'', True)

after u7d.send_signal(signal.SIGINT), and it's utterly important it is after,
othwrwise sessions are not properley terminated and the iptv vlan gets flooded.

This way is perfect, we add a little but bearable overhead using asuncio.wait_for()
to have quick detection of the end of a strean, without it it can take as much as
30 or 40s, and the client quickly requests the video from that moment on instead
of requesting again the same video,_this_ difference means the world, we don't have
to face an endless loop of the same stream again and again, there is minimal glitch
in the video and continues at the next second, so beautiful.

On the other hand, backtrack the relevant code in /app/u7d.py to try to switch
to the next program, it was a hack with ugly results.

Signed-off-by: Javier Marcet <javier@marcet.info>