movistar-u7d.py & movistar-epg.py: generalize catchup url format for broader compatibility.
Now the last part of the url must only comply with:

{timestamp}.{extension}

The timestamp must be a 10 digits string. Optionally it can also specfify a duration:

{timestamp}-{duration}.{extension}

The duration, when it exists, must be 1 or more digits.

Additionaly, it accepts a leading word separated by a dash, so it
accetps TiviMate's format:

video-{timestamp}-{duration}.{extension}

OTT Video Player uses:

timeshift_abs-{timestamp}.{extension}

Signed-off-by: Javier Marcet <javier@marcet.info>