movistar_u7d.py: fix stale vod.py processes launched for recordings once the recordings are done. The reason this happened is movistar_u7d needs to launch said processes as background processes or it will be forced to keep the record api calls open for the duration of the recordings.
By default, posix mandates parent processes must take care of their
children, i.e., they must wait() for them. When a child ends, the
kernel sends a SIG_CHLD signal so the parent knows and can wait() for
it. But there exists also the option to tell the kernel up front we
are careless parents and to not expect us to take care of our children
which is what we want movistar_u7d to do. vod.py is its only child,
even if repeated, it's always the same child and movistar_u7d does not
care at all about it when it is asked to record something, vod.py can
take care of itself on its own.

Hence, movistar_u7d.py now ignores the SIG_CHLD signal each time a vod
recording process is launched.

Signed-off-by: Javier Marcet <javier@marcet.info>