Control dynamically the number of allowed clients (and recordings), although only for installations within the router itself, since to control the iptv bw, the proxy needs direct access to the iptv interface, otherwise measuring the used bw is a much more complicated matter.
The maximum allowed iptv bw is set to a sane default of 85000 kbps,
out of a maximum capacity of 100000 kbits/second.

Since the avergae HD channel is around 6-8mbps, 85000 should be enough
to avoid deadlocking the remaining clients. This limit can be adjusted
with the env var IPTV_BW with the allowed bw in kbps, up to a maximum
of 90000, anything above will be cut down to 90000 kbps.

This is to ensure the smooth operation of all the existing clients,
when the limit if exceeded, streams start to drop packets.

Last, to also ensure regular clients are not rejected due to ongoing
recordings, with the IPTV_BW_RECS, by default set at 75000, you can
set the maximum bw usage over which recordings will be rejected.

It might me interesting to cancel the most recent ongoing recording
when a regular client wants to tune open a stream, I have to measure
if it would be fas enough to still be seamless for regular clients,
i.e., not seeing a 503 upfront and forcing the user (if the client is
not smart emough) to reopen the stream.

Signed-off-by: Javier Marcet <javier@marcet.info>