movistar_u7d.py, vod.py: turns out prioritizing normal clients over recordings by killing one of the later, works quite smoothly.
This proxy is now able to serve cathcup programs as fast as Movistar
can deliver them, whilt at the same time maximizing the available
iptv bandwidth to server multiple clients and handle simultaneous
recordings, all without disturbing ongoing normal clients at any
moment.

Recordings are repeated when they fail so no issue killing them when
the need arises. Also bear in mind the memory allocated to the
container if handling recordings dynamically. 2GB seems to work
quite well to consume the 100mbps of iptv bw.

Besides that, the server is also able to cope with ddos attacks quite
well, all, as usual, without ongoing clients noticing anything.

In other words, it is ready to serve the bigggest of the families who
might be Movistar Fiber IPTV's customers.

TV glitches with Movistar IPTV (with decent network hardware) are a
thing of the past :D

Signed-off-by: Javier Marcet <javier@marcet.info>