movistar_epg.py: add a new recents category to exported /recordings.m3u with the 100 most recent video files found within RECORDINGS, ordered from newest to oldest.
This makes it really easy to identiy and locate recordings as they're
added to the collection, as well as accessing very quickly to the
newest recordings.

When the the recorings collection has some size it is hard to realize
in what group (path) new recorings have been added. It is shown as
the top most folder in TiviMate's Movies section.

Signed-off-by: Javier Marcet <javier@marcet.info>