Add new module netifaces and new optional env var IPTV_IFACE to be able to choose which IPTV interface we want the proxy to run on and where we might also want to set IPTV_BW. This only applies to unix.
Geared specially to run the proxy within the router itself, where,
at boot time, the interfaces can take a variable amount of time to
become available and routed, although it might be useful in any case
where we have more than one interface in the system.

This improves the former hackerish way to determine the iptv
interface.

Signed-off-by: Javier Marcet <javier@marcet.info>