movistar_u7d.py & vod.py: finally real proper handling of recording processes on unix.
No more vod zombies and as an added bonus, in case either ffmpeg or
vod die, however badly it happens, the event is properly signaled
to the epg backend so the recording is removed from the queue and,
a new /timers_check is queued.

Thus, when either ffmpeg or vod itself die, it is all cleaned up in
a proper way and the recording requeud immediately.

Signed-off-by: Javier Marcet <javier@marcet.info>