movistar_u7d.py & vod.py: the signal handling for forken children turned out not to be enough. SIGCHLD was being received only for a few seconds after the child was launched, so it was neccesary to add a task for each children to cleanup after them.
With this, bandwidth consumption of vod processes is always stopped
the instant either ffmpeg or the vod process die. Only SIGKILL is
not trapped, which cannot. This also applies to Windows systems.

The signal handling is completely substituted by the manual cleanup,
and far from taxing the system too much, it performs amazingly well.

All this still works with movistar_u7d working in multiple processes.

And last, the cleanup queues a new /timers_check after a child ends,
unless the network is saturated, so the queue of recordings can
always be using the entire bandwidth available, non stop :D If a
normal client tries to open a stream while this happens, a recording
is nicely killed and requeued when there is available bw again.

Signed-off-by: Javier Marcet <javier@marcet.info>