movistar_epg.py & co: backtrack the 'smartness' figuring out premieres from reruns and instead let the user choose at what time a timer must match. It is flexible enough to match epg entries within 15 minutes of the specified time in timers.conf
To specify a time, a new field is added to the timers. As it is done
to choose the default language, separated with " ## ", now we can
also specify a HH:MM slot, like this:

4714 = [
    "American Dad ## 13:00",
    "American Dad ## 14:00 ## VO",
    "Los Simpson ## Esp ## 13:45",
]

The order of the fields for language and time does not matter. If we
want to match several times of the day we can repeat the timer, like
above.

In case we repeat the same field, like:

    "American Dad ## Esp ## VO"

or

    "American Dad ## 12:15 ## 15:00"

the last option will overwrite the first one.

Last, if we add more fields, like:

    "American Dad ## Esp ## 12:45 ## VO ## 15:00"

only the first two after the timer itself will be taken into account,
everything else will we ignored, so in the above case it will record
programs which air around 12:45 with Spanish as main audio track.

Signed-off-by: Javier Marcet <javier@marcet.info>