vod.py & co: new comskip postprocess step. It takes some time to process recordings but it is absolutely worth the wait. The included tuned config does quite a remarkable job marking commercials, in most cases it has an amazing accuracy, although it sure can be improved.
Both mpv & vlc make it really easy to skip chapters, Kodi can do it
automatically when it has access to the filesystem directly and can
read the '.vdr' files produced by comskip.

It is probably too expensive to run on smaller devices, but it is opt
in, i.e., it has to be enabled with the COMSKIP env var.

Signed-off-by: Javier Marcet <javier@marcet.info>