movistar_u7d: find simple & elegant solution which solves a year old bug filed at the Sanic project.
When the server exits, Sanic seems to not be cancelling & awaiting
tasks properly. With added code to the before_server_stop, the
crucial tasks are cancelled properly, hence they do their cleanup,
which in this case means stopping a network connection which would
be consuming bandwith for up to 60s more if not cancelled.

Signed-off-by: Javier Marcet <javier@marcet.info>